+++ samples/HostVirtualNetwork/Program.cs
@@ -94,6 +94,10 @@ if (!graphOnly)
 {
     cloudShell.AddApplicationProvider();
 }
+else
+{
+    cloudShell.AddApplicationResourceManagerUi();
+}
 
 cloudShell.Resources(resources =>
 {