+++ samples/SettingsAndSecrets/Program.cs
@@ -176,6 +176,10 @@ if (!graphOnly)
             options.ServiceAuthenticationSigningKeyPem = identitySigningKeyPem;
         });
 }
+else
+{
+    cloudShell.AddApplicationResourceManagerUi();
+}
 
 cloudShell.Resources(resources =>
 {